FIX hour sever
The real problem was in frontend, 'moment' directive with the method
format add or substract automatically time depend in what computer time
you are executing the method. As this server is in UTC and my software is
in UTC-5 I need really substract it. After I need to do it dynamically
because my clients will be of differents countries.